+++ src/main/java/roomescape/member/config/WebMvcConfiguration.java
@@ -1,11 +1,13 @@
-package roomescape.member;
+package roomescape.member.config;
 
 import java.util.List;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import roomescape.member.AuthService;
 import roomescape.member.service.MemberService;
 
 @Configuration
@@ -16,6 +18,12 @@ public class WebMvcConfiguration implements WebMvcConfigurer {
     public WebMvcConfiguration(AuthService authService, MemberService memberService) {
         this.authService = authService;
         this.memberService = memberService;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new CheckRoleInterceptor(authService, memberService))
+                .addPathPatterns("/admin/**");
     }
 
     @Override
